pyproject.toml
[project]
name = "tramontane"
version = "0.2.3"
description = "Mistral-native agent orchestration framework — EU sovereign, GDPR built-in"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    {name = "Jesiel", email = "rombley.jesiel@gmail.com"},
]
keywords = [
    "mistral", "agents", "orchestration", "llm",
    "pipeline", "gdpr", "eu-sovereign", "ai-framework",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Framework :: FastAPI",
    "Typing :: Typed",
]

dependencies = [
    "mistralai>=1.0.0",
    "pydantic>=2.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "anyio>=4.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "huggingface-hub>=1.8.0",
]

[project.optional-dependencies]
redis = ["redis>=5.0.0"]
postgres = ["asyncpg>=0.29.0", "pgvector>=0.2.0"]
voice = ["sounddevice>=0.4.0", "numpy>=1.26.0"]
sandbox = ["e2b>=0.17.0"]
hub = ["huggingface-hub>=0.20.0"]

[project.scripts]
tramontane = "tramontane.cli.main:app"

[project.urls]
Homepage = "https://github.com/Jesiel-dev-creator/TRAMONTANE"
Repository = "https://github.com/Jesiel-dev-creator/TRAMONTANE"
Documentation = "https://github.com/Jesiel-dev-creator/TRAMONTANE#readme"
"Bug Tracker" = "https://github.com/Jesiel-dev-creator/TRAMONTANE/issues"
"HuggingFace Demo" = "https://huggingface.co/spaces/BleuCommerce-Apps/TRAMONTANE-demo"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
    "types-pyyaml>=6.0.12.20250915",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
